# functions/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This API is the easiest way to interact with the WeaverFi SDK. Find out more on [GitHub](https://github.com/WeaverFi/api), or join our [Discord Server](https://discord.com/invite/DzADcq7y75).",
    "version": "1.10.2",
    "title": "WeaverFi API",
    "license": {
      "name": "MIT License",
      "url": "https://github.com/WeaverFi/api/blob/main/LICENSE.md"
    }
  },
  "schemes": [
    "https"
  ],
...
  "paths": {
    "/chains": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets all blockchains supported by WeaverFi.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/Chains"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/projects": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets all projects supported by WeaverFi.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/Projects"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/tokens": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets all tokens automatically tracked by WeaverFi.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/Tokens"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/info": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets information regarding any individual blockchain.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/ChainInfo"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/projects": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets projects supported on any individual blockchain.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/ChainProjects"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/tokens": {
      "get": {
        "tags": [
          "Info"
        ],
        "summary": "Gets tokens automatically tracked on any individual blockchain.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/ChainTokens"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/tokenPrices": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Returns pricing information of all tracked tokens.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/TokenPrices"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/nativeTokenPrices": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Returns pricing information of all chain's native tokens.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/TokenPrices"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/tokenPrices": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Returns pricing information of tracked tokens on any individual blockchain.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/ChainTokenPrices"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/tokenPrice": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Returns pricing information of a specific token.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "address",
            "in": "query",
            "description": "Token address.",
            "required": true,
            "type": "string"
          },
          {
            "name": "decimals",
            "in": "query",
            "description": "Token decimals.",
            "type": "integer"
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/TokenPrice"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/tokenPriceHistory": {
      "get": {
        "tags": [
          "Prices"
        ],
        "summary": "Returns price history of a specific token.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "address",
            "in": "query",
            "description": "Token address.",
            "required": true,
            "type": "string"
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/TokenPriceHistory"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/wallet": {
      "get": {
        "tags": [
          "Balances"
        ],
        "summary": "Gets token balances in any wallet address.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "address",
            "in": "query",
            "description": "Wallet address.",
            "required": true,
            "type": "string"
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/WalletTokenBalances"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
    "/{chain}/project": {
      "get": {
        "tags": [
          "Balances"
        ],
        "summary": "Gets staked token balances in any project, by any wallet address.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "name",
            "in": "query",
            "description": "Project name. (See /projects for valid project names)",
            "required": true,
            "type": "string"
          },
          {
            "name": "address",
            "in": "query",
            "description": "Wallet address.",
            "required": true,
            "type": "string"
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/ProjectTokenBalances"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
...
    "/{chain}/nfts": {
      "get": {
        "tags": [
          "Balances"
        ],
        "summary": "Gets NFT balances in any wallet address.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "chain",
            "in": "path",
            "description": "Blockchain to query.",
            "required": true,
            "type": "string",
            "enum": [
              "eth",
              "bsc",
              "poly",
              "ftm",
              "avax",
              "cronos",
              "op",
              "arb"
            ]
          },
          {
            "name": "address",
            "in": "query",
            "description": "Wallet address.",
            "required": true,
            "type": "string"
          },
          {
            "name": "key",
            "in": "query",
            "description": "Your API key.",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful query.",
            "schema": {
              "$ref": "#/definitions/WalletNFTBalances"
            }
          },
          "400": {
            "description": "Bad request."
          },
          "401": {
            "description": "Invalid API key."
          },
          "429": {
            "description": "Too many requests (rate limited)."
          },
          "500": {
            "description": "Internal API error."
          }
        }
      }
    },
...
  },
  "definitions": {
    "Chains": {
      "type": "object",
      "properties": {
        "chain": {
          "$ref": "#/definitions/ChainInfo"
        }
      }
    },
    "Projects": {
      "type": "object",
      "properties": {
        "chain": {
          "$ref": "#/definitions/ChainProjects"
        }
      }
    },
    "Tokens": {
      "type": "object",
      "properties": {
        "chain": {
          "$ref": "#/definitions/ChainTokens"
        }
      }
    },
...
    "ChainProjects": {
      "type": "array",
      "items": {
        "type": "string"
      },
      "example": [
        "aave",
        "curve",
        "pooltogether"
      ]
    },
    "ChainTokens": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/Token"
      }
    },
...
    "TokenPrices": {
      "type": "object",
      "properties": {
        "chain": {
          "$ref": "#/definitions/ChainTokenPrices"
        }
      }
    },
    "ChainTokenPrices": {
      "type": "array",
      "items": {
        "$ref": "#/definitions/TokenPriceData"
      }
    },
...
    "TokenPrice": {
      "type": "object",
      "properties": {
        "chain": {
          "type": "string",
          "example": "ETH"
        },
        "address": {
          "type": "string",
          "example": "0xeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeeee"
        },
        "decimals": {
          "type": "integer",
          "example": 18
        },
        "price": {
          "type": "integer",
          "example": 2890.76
        }
      }
    },
...
  },
...
}